Rename auth to gateway
