+++ gateway/cdk.ts
