+++ gateway/lambdas/create-auth-challenge.ts
