+++ gateway/lambdas/define-auth-challenge.ts
