+++ gateway/lambdas/pre-sign-up.ts
