+++ gateway/lambdas/verify-auth-challenge-response.ts
