+++ gateway/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@internote/auth-service",
+  "name": "@internote/gateway-service",
   "version": "1.0.0",
   "description": "",
   "author": "",
