+++ package.json
@@ -13,7 +13,7 @@
   "workspaces": {
     "packages": [
       "attachments",
-      "auth",
+      "gateway",
       "collaborate",
       "dictionary",
       "export",
